refactor(admin): fold the Illuminate route into the CLI explorer
Closes #651.

The standalone `/illuminate` route duplicated graph exploration that the CLI
explorer (`/cli`) already does better: the CLI owns command history, the axis
picker, the splitter, and canvas persistence. This removes the duplicate
surface and makes the CLI canvas the single place to visualise a
neighbourhood.

What changed:

- Delete the `/illuminate` route and its page-accumulator UI (seed prompt,
  toolbar, neighbour table, lineage chips, node-detail panel). The reusable
  `IlluminateCanvas` component is KEPT — it now renders only inside the CLI
  canvas panel.
- Re-point every entry affordance at `/cli?seed=<key>`: the Browse Vertices /
  Browse Edges row buttons and the vertex-detail toolbar. `CliPage` consumes
  `?seed=` once on mount and dispatches the canonical click form
  `illuminate <seed> 2 5`, matching what a canvas node click already runs.
- Drop the Illuminate nav item and the landing FEATURES card.
- Prune `usecase/illuminate` to the canvas-facing selectors/state only;
  delete the now-orphaned `vertex-picker` usecase and `VertexPicker`
  component (their sole consumers were the deleted SeedPrompt + toolbar).
- E2E: re-point the browse/landing navigation specs at `/cli?seed=`; rename
  `illuminate.spec.ts` to `canvas.spec.ts`, porting the three pure-canvas
  regression guards (WCAG AA label contrast #453, label toggles #517,
  drag-release-without-pin #491) onto the `illuminate <hub> 2 5` bootstrap,
  which reproduces the old initial 2-hop frame. The accumulator/expansion
  -coupled guards that no longer have a UI are tracked for CLI-native
  restoration in #664.
