--- admin/app/components/illuminate/IlluminatePage/IlluminatePage.module.css
@@ -1,40 +1,0 @@
-.root {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacingVerticalL, 16px);
-}
-
-.header {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacingVerticalS, 8px);
-}
-
-.title {
-  font-size: var(--fontSizeBase600, 24px);
-  font-weight: var(--fontWeightSemibold, 600);
-  margin: 0;
-}
-
-.lead {
-  color: var(--colorNeutralForeground2, #424242);
-  margin: 0;
-}
-
-.alert {
-  margin: 0;
-}
-
-.tableDisclosure {
-  border: 1px solid var(--colorNeutralStroke2, #e0e0e0);
-  border-radius: var(--borderRadiusMedium, 6px);
-  background: var(--colorNeutralBackground1, #fff);
-  padding: var(--spacingVerticalS, 8px) var(--spacingHorizontalM, 12px);
-}
-
-.summary {
-  cursor: pointer;
-  color: var(--colorNeutralForeground2, #424242);
-  font-size: var(--fontSizeBase200, 12px);
-  padding: 4px 0;
-}
