--- admin/app/components/illuminate/IlluminateToolbar/IlluminateToolbar.module.css
@@ -1,79 +1,0 @@
-.toolbar {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacingVerticalM, 12px);
-  padding: var(--spacingVerticalM, 12px) var(--spacingHorizontalM, 12px);
-  border: 1px solid var(--colorNeutralStroke2, #e0e0e0);
-  border-radius: var(--borderRadiusMedium, 6px);
-  background: var(--colorNeutralBackground1, #fff);
-}
-
-.seedRow {
-  display: flex;
-  align-items: center;
-  gap: var(--spacingHorizontalS, 8px);
-  flex-wrap: wrap;
-}
-
-.seedLabel {
-  color: var(--colorNeutralForeground2, #424242);
-  font-size: var(--fontSizeBase200, 12px);
-}
-
-.seedValue {
-  flex: 1 1 240px;
-  min-width: 0;
-  padding: 4px 8px;
-  background: var(--colorNeutralBackground3, #f5f5f5);
-  border-radius: var(--borderRadiusMedium, 6px);
-  color: var(--colorNeutralForeground1, #242424);
-  font-family: var(
-    --fontFamilyMonospace,
-    ui-monospace,
-    SFMono-Regular,
-    monospace
-  );
-  font-size: var(--fontSizeBase200, 12px);
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.chipStripSlot {
-  flex: 1 1 240px;
-  min-width: 0;
-}
-
-.expandRow {
-  display: flex;
-  align-items: flex-end;
-  gap: var(--spacingHorizontalS, 8px);
-  flex-wrap: wrap;
-}
-
-.expandPickerSlot {
-  flex: 1 1 320px;
-  min-width: 0;
-  max-width: 480px;
-}
-
-.knobs {
-  display: flex;
-  align-items: center;
-  gap: var(--spacingHorizontalL, 16px);
-  flex-wrap: wrap;
-}
-
-.knobField {
-  display: flex;
-  align-items: center;
-  gap: var(--spacingHorizontalS, 8px);
-  min-width: 220px;
-  flex: 1 1 220px;
-}
-
-.optimization {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacingVerticalXS, 4px);
-}
