--- admin/app/components/illuminate/SeedPrompt/SeedPrompt.module.css
@@ -1,10 +1,0 @@
-.form {
-  display: flex;
-  flex-direction: column;
-  gap: var(--spacingVerticalM, 12px);
-  max-width: 520px;
-  padding: var(--spacingVerticalL, 16px);
-  border: 1px solid var(--colorNeutralStroke2, #e0e0e0);
-  border-radius: var(--borderRadiusMedium, 6px);
-  background: var(--colorNeutralBackground1, #fff);
-}
