+++ admin/app/lib/client/usecase/illuminate/selectors.ts
@@ -1,24 +1,4 @@
 import type { Edge, Vertex } from "~/lib/client/infrastructure/api/illuminate";
-import { ACCUMULATOR_SOFT_CAP, type IlluminateState } from "./state";
-
-/**
- * `true` when the given protobuf-JSON timestamp is at or before the
- * supplied wall clock. Missing / unparseable timestamps are treated as
- * "never expires" (returns `false`) — defensive: we'd rather render a
- * value at full opacity than drop it over a malformed ISO string.
- *
- * Mirrors the semantics of
- * `IlluminateCanvas/ttl-decay.ts:computeTtlFraction` but specialised
- * to the boolean "should this be filtered?" question the selector
- * cares about. Kept here (not imported from the canvas package) so
- * the use-case layer stays free of component-side dependencies.
- */
-function isExpired(expiration: string | undefined, nowMs: number): boolean {
-  if (expiration === undefined || expiration === "") return false;
-  const expiresAtMs = Date.parse(expiration);
-  if (!Number.isFinite(expiresAtMs)) return false;
-  return expiresAtMs <= nowMs;
-}
 
 /**
  * Graph node shape consumed by `IlluminateCanvas`. We keep the full
@@ -298,277 +278,4 @@ export function selectHopBucketCounts(
     { key: "far", count: far },
     { key: "unreachable", count: unreachable },
   ];
-}
-
-/**
- * Builds the view model the canvas renders from the accumulator. Pure;
- * no React, no DOM, no graphology — those live in the component layer
- * so this stays trivial to unit-test.
- *
- * `nowMs` is an injected "wall clock" used to drop already-expired
- * vertices and (cascading) their incident edges — per #459 the canvas
- * should never render a value that the server has already let go,
- * even if a stale response is still in the accumulator. The parameter
- * is optional and defaults to `Date.now()` because the vast majority
- * of test fixtures construct vertices without expiration; only the
- * decay-aware tests pass an explicit value.
- *
- * Mid-frame fading (per-tick alpha as a vertex approaches the
- * cliff) is NOT a selector concern — the canvas reducer reads the
- * expiration off each node and fades it on every refresh tick so we
- * avoid recomputing the whole view model every second. See
- * {@link IlluminateCanvas} `nowRef` / `tickRef` for the per-frame
- * machinery.
- */
-export function selectGraphView(
-  state: IlluminateState,
-  nowMs: number = Date.now(),
-): GraphView {
-  const initialSeed = state.initialSeed;
-  const expansionOrigins = state.expansions.map((e) => e.origin);
-  const latestExpansionOrigin =
-    expansionOrigins.length > 0
-      ? (expansionOrigins[expansionOrigins.length - 1] ?? null)
-      : null;
-
-  // #483: the "latest result" is the membership of the most recent
-  // expansion. The canvas hides everything outside it. Empty sets when
-  // there are no expansions act as "no filter" downstream.
-  const latestExpansion =
-    state.expansions.length > 0
-      ? state.expansions[state.expansions.length - 1]
-      : undefined;
-  const latestResultVertexKeys = new Set(latestExpansion?.vertexKeys ?? []);
-  const latestResultEdgeIds = new Set(latestExpansion?.edgeIds ?? []);
-
-  if (state.accumulator.vertices.size === 0) {
-    return {
-      nodes: [],
-      edges: [],
-      latestExpansionOrigin,
-      expansionOrigins,
-      overSoftCap: false,
-      latestResultVertexKeys,
-      latestResultEdgeIds,
-    };
-  }
-
-  const originSet = new Set(expansionOrigins);
-
-  // Edge weights drive node importance. Aggregate over the merged edges,
-  // skipping ones that are already past expiry — a dying edge shouldn't
-  // inflate a node's importance score (#459).
-  const weightByKey = new Map<string, number>();
-  for (const acc of state.accumulator.edges.values()) {
-    const e = acc.edge;
-    if (!e.tail || !e.head) continue;
-    if (isExpired(e.expiration, nowMs)) continue;
-    const w = e.weight ?? 0;
-    weightByKey.set(e.tail, (weightByKey.get(e.tail) ?? 0) + w);
-    weightByKey.set(e.head, (weightByKey.get(e.head) ?? 0) + w);
-  }
-  const maxWeight = Math.max(0, ...weightByKey.values());
-
-  const nodes: GraphNode[] = [];
-  for (const [key, acc] of state.accumulator.vertices) {
-    if (key === "") continue;
-    // #459: filter expired vertices at selector time. The next fetch
-    // would drop them anyway, but explicit removal here means a stale
-    // accumulator never renders a tombstone.
-    if (isExpired(acc.vertex.expiration, nowMs)) continue;
-    const isInitialSeed = initialSeed !== null && key === initialSeed;
-    const isExpansionOrigin = originSet.has(key);
-    const raw = weightByKey.get(key) ?? 0;
-    const importance =
-      isInitialSeed || isExpansionOrigin
-        ? 1
-        : maxWeight > 0
-          ? Math.max(0.1, raw / maxWeight)
-          : 0.5;
-    const firstSeenExpansion = acc.expansionIndexes[0] ?? 0;
-    nodes.push({
-      id: key,
-      label: key,
-      vertex: acc.vertex,
-      isInitialSeed,
-      isExpansionOrigin,
-      importance,
-      firstSeenExpansion,
-      // Placeholder; filled in below after edges are filtered so the
-      // BFS adjacency mirrors the live (non-expired) graph.
-      hopDistance: Number.POSITIVE_INFINITY,
-    });
-  }
-
-  const knownKeys = new Set(nodes.map((n) => n.id));
-  const edges: GraphEdge[] = [];
-  for (const [id, acc] of state.accumulator.edges) {
-    const e = acc.edge;
-    if (!e.tail || !e.head) continue;
-    // Drop edges that reference vertices we don't have; the canvas would
-    // otherwise throw when sigma tries to look them up. (Shouldn't happen
-    // with current server behaviour, but defensive.) This also handles
-    // the #459 cascade where an expired vertex was filtered above.
-    if (!knownKeys.has(e.tail) || !knownKeys.has(e.head)) continue;
-    // #459: an edge can have its own TTL that's shorter than either
-    // endpoint's. Filter past-expiry edges explicitly.
-    if (isExpired(e.expiration, nowMs)) continue;
-    edges.push({
-      id,
-      source: e.tail,
-      target: e.head,
-      weight: e.weight ?? 0,
-      edge: e,
-    });
-  }
-
-  // #460: multi-source BFS over the undirected projection of the live
-  // (post-filter) edge set, seeded by every expansion origin that is
-  // still in the accumulator. Vertices unreachable from every origin
-  // keep the `Number.POSITIVE_INFINITY` placeholder so the canvas
-  // reducer can render them with the "unreachable" tone.
-  const hopByKey = computeHopDistances(knownKeys, edges, expansionOrigins);
-  for (const node of nodes) {
-    node.hopDistance = hopByKey.get(node.id) ?? Number.POSITIVE_INFINITY;
-  }
-
-  return {
-    nodes,
-    edges,
-    latestExpansionOrigin,
-    expansionOrigins,
-    overSoftCap: state.accumulator.vertices.size > ACCUMULATOR_SOFT_CAP,
-    latestResultVertexKeys,
-    latestResultEdgeIds,
-  };
-}
-
-export function selectIsBusy(state: IlluminateState): boolean {
-  return state.status === "loading" || state.pendingCount > 0;
-}
-
-export function selectExpansionCount(state: IlluminateState): number {
-  return state.expansions.length;
-}
-
-export function selectCanClear(state: IlluminateState): boolean {
-  return state.expansions.length > 0 || state.accumulator.vertices.size > 0;
-}
-
-/**
- * One outgoing edge as surfaced by the node-detail Drawer (#461). The
- * `id` mirrors `edgeIdOf(tail, head)` so React keys stay stable and the
- * caller can cross-reference the accumulator if needed.
- */
-export interface OutgoingEdge {
-  id: string;
-  /** Head vertex key (the edge's destination). */
-  target: string;
-  weight: number;
-  /** Proto-JSON expiration of the edge, if any. */
-  expiration?: string;
-}
-
-/**
- * View-model the node-detail Drawer renders for the inspected vertex
- * (#461). Built purely from the accumulator so opening the Drawer never
- * issues an RPC — outgoing edges come straight from
- * `state.accumulator.edges`. Inbound edges are NOT included here: they
- * require a fresh `ScanEdges({ headPrefix })` and are fetched on demand
- * by {@link useInboundEdges} into panel-local state so they never
- * mutate the canvas accumulator (#466 keeps the canvas the sole graph
- * owner).
- */
-export interface InspectedVertexDetail {
-  key: string;
-  vertex: Vertex;
-  /**
-   * Live (non-expired) outgoing edges whose tail is `key`, sorted by
-   * target key for a stable render order.
-   */
-  outgoing: OutgoingEdge[];
-}
-
-/**
- * Builds the {@link InspectedVertexDetail} for the supplied key, or
- * `null` when the vertex is absent / expired. Returning `null` for a
- * vanished vertex lets the page drive the Drawer's `open` state straight
- * off this selector: if a TTL sweep drops the inspected vertex, the
- * Drawer self-closes on the next render without any extra wiring.
- *
- * Pure (no React, no client) so the accumulator → panel projection is
- * unit-testable in isolation.
- */
-export function selectInspectedDetail(
-  state: IlluminateState,
-  key: string | null,
-  nowMs: number = Date.now(),
-): InspectedVertexDetail | null {
-  if (key === null || key === "") return null;
-  const acc = state.accumulator.vertices.get(key);
-  if (!acc) return null;
-  if (isExpired(acc.vertex.expiration, nowMs)) return null;
-
-  const outgoing: OutgoingEdge[] = [];
-  for (const [id, accEdge] of state.accumulator.edges) {
-    const e = accEdge.edge;
-    if (e.tail !== key || !e.head) continue;
-    if (isExpired(e.expiration, nowMs)) continue;
-    outgoing.push({
-      id,
-      target: e.head,
-      weight: e.weight ?? 0,
-      expiration: e.expiration,
-    });
-  }
-  outgoing.sort((a, b) =>
-    a.target < b.target ? -1 : a.target > b.target ? 1 : 0,
-  );
-
-  return { key, vertex: acc.vertex, outgoing };
-}
-
-/**
- * Narrows a `ScanEdges({ headPrefix: key })` result to the edges that
- * terminate at exactly `key` (#461 "Show inbound edges"). The wire
- * `headPrefix` scan is a prefix match, so `key = "a"` would also return
- * `a*→…` edges whose head merely starts with the prefix; the Drawer
- * only wants true inbound edges, hence the exact `head === key` filter.
- * Pure so it can be unit-tested without a live client.
- */
-export function filterInboundEdges(edges: Edge[], key: string): Edge[] {
-  if (key === "") return [];
-  return edges.filter((e) => e.head === key);
-}
-
-/**
- * View-model for the per-expansion chip in the toolbar (#456). Each chip
- * is a "scroll-to-origin" affordance — clicking it pans the camera to the
- * origin vertex without mutating state or re-issuing any RPC.
- *
- * Order matches `state.expansions` (chronological); `chips[0]` is the
- * seed expansion (#466 D5) and carries `isSeed: true` so the toolbar
- * can render a star/badge marker. Duplicates by `originKey` are
- * preserved on purpose: clicking the same vertex twice produces two
- * chips with the same label because the user genuinely performed two
- * exploration actions and may want to revisit either point in time.
- */
-export interface ExpansionChip {
-  /** Stable React key — mirrors `Expansion.id`. */
-  id: number;
-  /** Display label + camera target. */
-  originKey: string;
-  /** True only for `expansions[0]` (the URL-level seed). */
-  isSeed: boolean;
-  /** Zero-based position in `state.expansions` for tooltips. */
-  index: number;
-}
-
-export function selectExpansionChips(state: IlluminateState): ExpansionChip[] {
-  return state.expansions.map((exp, index) => ({
-    id: exp.id,
-    originKey: exp.origin,
-    isSeed: index === 0,
-    index,
-  }));
 }
